+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog
 ------
 
 - removed a minor bug in `nodes.py` that led to issues with retrieving variable indices from single-neuron nodes
+- removed a bug in `nodes.py` where, in some cases, the state vector `y` was not moved to the correct device after state resets
 
 0.10.2
 ------